+++ Java25-Team1-FinalProject/src/main/java/com/example/team_project/services/UserService.java
@@ -22,12 +22,12 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(Math.toIntExact(id));
+    public Optional<User> getUserById(Integer id) {
+        return userRepository.findById(id);
     }
 
-    public User updateUser(Long id, User userDetails) {
-        User user = userRepository.findById(Math.toIntExact(id))
+    public User updateUser(Integer id, User userDetails) {
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalStateException("User not found"));
 
         user.setFirstName(userDetails.getFirstName());
@@ -39,14 +39,14 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Long id) {
-        User user = userRepository.findById(Math.toIntExact(id))
+    public void deleteUser(Integer id) {
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalStateException("User not found"));
         userRepository.delete(user);
     }
 
-    public void deactivateUser(Long id) {
-        User user = userRepository.findById(Math.toIntExact(id))
+    public void deactivateUser(Integer id) {
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalStateException("User not found"));
         user.setActive(false);
         userRepository.save(user);
